Adding our game mod's documentation site to your list! :)
I'm the creator of a game mode for the military simulator Arma 3 called Invade & Annex. It's the first of its kind to introduce open source mission creation using the community to contribute towards one massive goal! It'd be an honour if you included it on your list; it's a fantastic piece of software you've created!