Resolve theme issues with https behind reverse proxy
Resolving 2 issues by changing the protocol to //.
When terminating https at the ACE level the HTTPS field is false
resulting in http requests for css on an https site.

Requesting the Roboto Slab font always goes over http. This is a problem
for browsers when the site is hosted over https.